Fix webhook signature verification to use the webhook signing secret (#351)
Since [the other PR](https://github.com/mailgun/mailgun-go/pull/321)
didn't seem to be being addressed, I'm submitting my own. The issues
brought up in it have been resolved. Tests pass and it seems to work as
directed when testing it as a `replace` in a different project. Please
let me know if anything needs addressed, I'd like to have this available
as soon as I can get it. Thanks!